+++ NWEEI/NWEEI/Views/Article/Edit.cshtml
@@ -19,25 +19,28 @@
         <form asp-action="Edit">
             <div asp-validation-summary="ModelOnly" class="text-danger"></div>
             <input type="hidden" asp-for="ArticleID" />
+            <input asp-for="DateCreated" type="hidden" value="@Model.DateCreated" />
+            <input asp-for="PublishDate" type="hidden" value="@Model.PublishDate" />
             <div class="form-group">
                 <label asp-for="Title" class="control-label"></label>
                 <input asp-for="Title" class="form-control" />
                 <span asp-validation-for="Title" class="text-danger"></span>
             </div>
             <div class="form-group">
+                <link rel="stylesheet" href="/richtexteditor/rte_theme_default.css" />
+                <script type="text/javascript" src="/richtexteditor/rte.js"></script>
+                <script type="text/javascript" src='/richtexteditor/plugins/all_plugins.js'></script>
+                <script type="text/javascript" src="/rte-upload.js"></script>
                 <label asp-for="Body" class="control-label"></label>
-                <input asp-for="Body" class="form-control" />
+                <input asp-for="Body" name="htmlcode" id="inp_htmlcode" type="hidden" class="form-control" />
+                <div id="div_editor1" class="richtexteditor" style="width: 960px;margin:0 auto;"></div>
+                <script>
+                    var editor1 = new RichTextEditor(document.getElementById("div_editor1"));
+                    editor1.attachEvent("change", function () {
+                        document.getElementById("inp_htmlcode").value = editor1.getHTMLCode();
+                    });
+                </script>
                 <span asp-validation-for="Body" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <label asp-for="DateCreated" class="control-label"></label>
-                <input asp-for="DateCreated" class="form-control" />
-                <span asp-validation-for="DateCreated" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <label asp-for="PublishDate" class="control-label"></label>
-                <input asp-for="PublishDate" class="form-control" />
-                <span asp-validation-for="PublishDate" class="text-danger"></span>
             </div>
             <div class="form-group form-check">
                 <label class="form-check-label">
